third_party/WebKit/LayoutTests/svg/parser/whitespace-length-expected.txt

This is a testharness.js-based test.
FAIL Test <length> trailing garbage, value: -47assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10
FAIL Test <length> trailing garbage, value: +32assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10em
FAIL Test <length> trailing garbage, value: +32emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10ex
FAIL Test <length> trailing garbage, value: +32exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10px
FAIL Test <length> trailing garbage, value: +32pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10in
FAIL Test <length> trailing garbage, value: +32inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10cm
FAIL Test <length> trailing garbage, value: +32cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10mm
FAIL Test <length> trailing garbage, value: +32mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10pt
FAIL Test <length> trailing garbage, value: +32ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10pc
FAIL Test <length> trailing garbage, value: +32pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: -47%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: .1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: 0.35%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: 1e-10%
FAIL Test <length> trailing garbage, value: +32%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: +17E-1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: 17e+2%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10
FAIL Test <length> trailing garbage, value:  +32assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10em
FAIL Test <length> trailing garbage, value:  +32emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10ex
FAIL Test <length> trailing garbage, value:  +32exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10px
FAIL Test <length> trailing garbage, value:  +32pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10in
FAIL Test <length> trailing garbage, value:  +32inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10cm
FAIL Test <length> trailing garbage, value:  +32cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10mm
FAIL Test <length> trailing garbage, value:  +32mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10pt
FAIL Test <length> trailing garbage, value:  +32ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10pc
FAIL Test <length> trailing garbage, value:  +32pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:  -47%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:  .1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:  0.35%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:  1e-10%
FAIL Test <length> trailing garbage, value:  +32%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:  +17E-1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:  17e+2%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10
FAIL Test <length> trailing garbage, value:    +32assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10em
FAIL Test <length> trailing garbage, value:    +32emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10ex
FAIL Test <length> trailing garbage, value:    +32exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10px
FAIL Test <length> trailing garbage, value:    +32pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10in
FAIL Test <length> trailing garbage, value:    +32inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10cm
FAIL Test <length> trailing garbage, value:    +32cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10mm
FAIL Test <length> trailing garbage, value:    +32mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10pt
FAIL Test <length> trailing garbage, value:    +32ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10pc
FAIL Test <length> trailing garbage, value:    +32pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value:    -47%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value:    .1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value:    0.35%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value:    1e-10%
FAIL Test <length> trailing garbage, value:    +32%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value:    +17E-1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value:    17e+2%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10
FAIL Test <length> trailing garbage, value: \r\n\t +32assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10em
FAIL Test <length> trailing garbage, value: \r\n\t +32emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10ex
FAIL Test <length> trailing garbage, value: \r\n\t +32exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10px
FAIL Test <length> trailing garbage, value: \r\n\t +32pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10in
FAIL Test <length> trailing garbage, value: \r\n\t +32inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10cm
FAIL Test <length> trailing garbage, value: \r\n\t +32cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10mm
FAIL Test <length> trailing garbage, value: \r\n\t +32mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10pt
FAIL Test <length> trailing garbage, value: \r\n\t +32ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10pc
FAIL Test <length> trailing garbage, value: \r\n\t +32pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \r\n\t -47%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \r\n\t .1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \r\n\t 0.35%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \r\n\t 1e-10%
FAIL Test <length> trailing garbage, value: \r\n\t +32%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \r\n\t +17E-1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \r\n\t 17e+2%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10
FAIL Test <length> trailing garbage, value: \f+32assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10em
FAIL Test <length> trailing garbage, value: \f+32emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2emassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10ex
FAIL Test <length> trailing garbage, value: \f+32exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2exassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10px
FAIL Test <length> trailing garbage, value: \f+32pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2pxassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10in
FAIL Test <length> trailing garbage, value: \f+32inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2inassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10cm
FAIL Test <length> trailing garbage, value: \f+32cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2cmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10mm
FAIL Test <length> trailing garbage, value: \f+32mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2mmassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10pt
FAIL Test <length> trailing garbage, value: \f+32ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2ptassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10pc
FAIL Test <length> trailing garbage, value: \f+32pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2pcassert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
FAIL Test <length> trailing garbage, value: \f-47%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got -47
FAIL Test <length> trailing garbage, value: \f.1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.10000000149011612
FAIL Test <length> trailing garbage, value: \f0.35%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 0.3499999940395355
PASS Test <length> trailing garbage, value: \f1e-10%
FAIL Test <length> trailing garbage, value: \f+32%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 32
FAIL Test <length> trailing garbage, value: \f+17E-1%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1.7000000476837158
FAIL Test <length> trailing garbage, value: \f17e+2%assert_approx_equals: expected 0 +/- 5.960464477539063e-8 but got 1700
Harness: the test ran to completion.
